Implement 'completions' op from nrepl 0.8
This adds a completions handler that matches the nrepl spec (more
or less.) For Lua, it just borrows the existing logic of the
pre-existing complete op handler, wrapping it.

For Fennel we just use the ,complete command because it's better :)

In both cases, we're only returning the :candidate key here, and we
don't respect the :options key of incoming messages (which lets
caller request doc and arglists along with the candidates.) I figure
that's well enough though, at least in Emacs-land I don't think
anybody cares about those (or really :type either for that matter),
because :company-kind, :company-docsig, and :annotation-function
exist.

And outside of Emacs, I don't think I've ever seen an autocomplete
UI that tries to show documentation for all/multiple candidates at
the same time either.